refactor(storybook): extract Composition and Extension into a dedicated story
