+++ .storybook/locales/langs/en-US.json
@@ -114,7 +114,8 @@
   "storybook.stories.CoreApi.sample.conflict.items": "Field items",
   "storybook.stories.CoreApi.sample.conflict.values": "Field values",
   "storybook.stories.CoreApi.sample.conflict.meta": "Field meta",
-  "storybook.stories.CorePatterns.metaDescription": "Bundles the more easily overlooked but highly practical README topics together: string localization, autoLabel, enum composition, read-only structures, and Enum.extends.",
+  "storybook.stories.CoreCompositionAndExtension.metaDescription": "Shows how to compose multiple enums into a full workflow with raw(), then attach a shared transform method to every enum instance through Enum.extends.",
+  "storybook.stories.CorePatterns.metaDescription": "Bundles the more easily overlooked but highly practical README topics together: string localization with autoLabel, and one global template that simplifies labeling for every enum.",
   "storybook.stories.CorePatterns.localization.title": "String localization and autoLabel",
   "storybook.stories.CorePatterns.localization.description": "This approach also works outside React. With localize + labelPrefix, enum labels become resource keys instead of hard-coded display text.",
   "storybook.stories.CorePatterns.localization.highlights.localize": "localize",
@@ -129,26 +130,26 @@
   "storybook.stories.CorePatterns.localization.sample.draft": "Draft",
   "storybook.stories.CorePatterns.localization.sample.review": "In Review",
   "storybook.stories.CorePatterns.localization.sample.published": "Published",
-  "storybook.stories.CorePatterns.composition.title": "Enum composition and global extension",
-  "storybook.stories.CorePatterns.composition.description": "After merging multiple enums into a full workflow, you can also attach a shared transform method to every enum instance via Enum.extends.",
-  "storybook.stories.CorePatterns.composition.highlights.extends": "Enum.extends",
-  "storybook.stories.CorePatterns.composition.highlights.raw": "raw()",
-  "storybook.stories.CorePatterns.composition.highlights.merge": "Merge Enums",
-  "storybook.stories.CorePatterns.composition.highlights.readonly": "Read-only Structure",
-  "storybook.stories.CorePatterns.composition.section.result.title": "Composition result",
-  "storybook.stories.CorePatterns.composition.section.result.description": "Use raw() to recover the original definitions, then rebuild a new enum from them.",
-  "storybook.stories.CorePatterns.composition.field.frozenEnum": "Object.isFrozen(enum)",
-  "storybook.stories.CorePatterns.composition.field.frozenItems": "Object.isFrozen(items)",
-  "storybook.stories.CorePatterns.composition.field.allLabels": "All labels",
-  "storybook.stories.CorePatterns.composition.field.enumName": "Enum name",
-  "storybook.stories.CorePatterns.composition.section.extension.title": "Custom extension method",
-  "storybook.stories.CorePatterns.composition.section.extension.description": "This demo adds toBadgeMap() through Enum.extends and turns a single enum into status badge ready data.",
-  "storybook.stories.CorePatterns.composition.section.extension.codeTitle": "toBadgeMap() registration code",
-  "storybook.stories.CorePatterns.composition.sample.draft": "Draft",
-  "storybook.stories.CorePatterns.composition.sample.review": "In Review",
-  "storybook.stories.CorePatterns.composition.sample.published": "Published",
-  "storybook.stories.CorePatterns.composition.sample.archived": "Archived",
-  "storybook.stories.CorePatterns.composition.sample.fullFlow": "Full Content Flow",
+  "storybook.stories.CoreCompositionAndExtension.title": "Enum composition and global extension",
+  "storybook.stories.CoreCompositionAndExtension.description": "After merging multiple enums into a full workflow, you can also attach a shared transform method to every enum instance via Enum.extends.",
+  "storybook.stories.CoreCompositionAndExtension.highlights.extends": "Enum.extends",
+  "storybook.stories.CoreCompositionAndExtension.highlights.raw": "raw()",
+  "storybook.stories.CoreCompositionAndExtension.highlights.merge": "Merge Enums",
+  "storybook.stories.CoreCompositionAndExtension.highlights.readonly": "Read-only Structure",
+  "storybook.stories.CoreCompositionAndExtension.section.result.title": "Composition result",
+  "storybook.stories.CoreCompositionAndExtension.section.result.description": "Use raw() to recover the original definitions, then rebuild a new enum from them.",
+  "storybook.stories.CoreCompositionAndExtension.field.frozenEnum": "Object.isFrozen(enum)",
+  "storybook.stories.CoreCompositionAndExtension.field.frozenItems": "Object.isFrozen(items)",
+  "storybook.stories.CoreCompositionAndExtension.field.allLabels": "All labels",
+  "storybook.stories.CoreCompositionAndExtension.field.enumName": "Enum name",
+  "storybook.stories.CoreCompositionAndExtension.section.extension.title": "Custom extension method",
+  "storybook.stories.CoreCompositionAndExtension.section.extension.description": "This demo adds toBadgeMap() through Enum.extends and turns a single enum into status badge ready data.",
+  "storybook.stories.CoreCompositionAndExtension.section.extension.codeTitle": "toBadgeMap() registration code",
+  "storybook.stories.CoreCompositionAndExtension.sample.draft": "Draft",
+  "storybook.stories.CoreCompositionAndExtension.sample.review": "In Review",
+  "storybook.stories.CoreCompositionAndExtension.sample.published": "Published",
+  "storybook.stories.CoreCompositionAndExtension.sample.archived": "Archived",
+  "storybook.stories.CoreCompositionAndExtension.sample.fullFlow": "Full Content Flow",
   "storybook.stories.CorePatterns.templates.page.title": "Global templates: one config for every enum",
   "storybook.stories.CorePatterns.templates.page.description": "Set Enum.config.templates once, then every enum only needs a minimal init object plus options.name. Keys are derived automatically and localized through Enum.localize.",
   "storybook.stories.CorePatterns.templates.mode.flat": "Flat Mode",
