+++ .storybook/locales/langs/zh-CN.json
@@ -114,7 +114,8 @@
   "storybook.stories.CoreApi.sample.conflict.items": "字段 items",
   "storybook.stories.CoreApi.sample.conflict.values": "字段 values",
   "storybook.stories.CoreApi.sample.conflict.meta": "字段 meta",
-  "storybook.stories.CorePatterns.metaDescription": "这个页面把 README 中较容易被忽略但非常实用的能力放到一起：字符串本地化、autoLabel、枚举组合、只读结构，以及通过 Enum.extends 增加全局扩展方法。",
+  "storybook.stories.CoreCompositionAndExtension.metaDescription": "展示如何通过 raw() 把多个枚举组合成完整流程，并通过 Enum.extends 将统一的转换方法挂载到所有 enum 实例上。",
+  "storybook.stories.CorePatterns.metaDescription": "本页把 README 中较容易被忽略但非常实用的能力放到一起：字符串本地化与 autoLabel，以及一次配置即可简化所有枚举标签的全局模板。",
   "storybook.stories.CorePatterns.localization.title": "字符串本地化与 autoLabel",
   "storybook.stories.CorePatterns.localization.description": "这是一种跨框架也适用的本地化方式。通过 localize + labelPrefix，可以把枚举项 label 变成资源 key，而不是写死展示文本。",
   "storybook.stories.CorePatterns.localization.highlights.localize": "localize",
@@ -129,26 +130,26 @@
   "storybook.stories.CorePatterns.localization.sample.draft": "草稿",
   "storybook.stories.CorePatterns.localization.sample.review": "审核中",
   "storybook.stories.CorePatterns.localization.sample.published": "已发布",
-  "storybook.stories.CorePatterns.composition.title": "枚举组合与全局扩展",
-  "storybook.stories.CorePatterns.composition.description": "把多个枚举合并成完整流程后，还可以通过 Enum.extends 把统一的数据变换方法挂到所有 enum 实例上。",
-  "storybook.stories.CorePatterns.composition.highlights.extends": "Enum.extends",
-  "storybook.stories.CorePatterns.composition.highlights.raw": "raw()",
-  "storybook.stories.CorePatterns.composition.highlights.merge": "合并枚举",
-  "storybook.stories.CorePatterns.composition.highlights.readonly": "只读结构",
-  "storybook.stories.CorePatterns.composition.section.result.title": "组合结果",
-  "storybook.stories.CorePatterns.composition.section.result.description": "利用 raw() 回收原始定义，再重新组合成新的 enum。",
-  "storybook.stories.CorePatterns.composition.field.frozenEnum": "Object.isFrozen(enum)",
-  "storybook.stories.CorePatterns.composition.field.frozenItems": "Object.isFrozen(items)",
-  "storybook.stories.CorePatterns.composition.field.allLabels": "所有 label",
-  "storybook.stories.CorePatterns.composition.field.enumName": "枚举名",
-  "storybook.stories.CorePatterns.composition.section.extension.title": "自定义扩展方法",
-  "storybook.stories.CorePatterns.composition.section.extension.description": "这里通过 Enum.extends 增加了 toBadgeMap()，把一份枚举统一转成适合状态徽标展示的数据结构。",
-  "storybook.stories.CorePatterns.composition.section.extension.codeTitle": "toBadgeMap() 注册代码",
-  "storybook.stories.CorePatterns.composition.sample.draft": "草稿",
-  "storybook.stories.CorePatterns.composition.sample.review": "审核中",
-  "storybook.stories.CorePatterns.composition.sample.published": "已发布",
-  "storybook.stories.CorePatterns.composition.sample.archived": "已归档",
-  "storybook.stories.CorePatterns.composition.sample.fullFlow": "完整内容流程",
+  "storybook.stories.CoreCompositionAndExtension.title": "枚举组合与全局扩展",
+  "storybook.stories.CoreCompositionAndExtension.description": "把多个枚举合并成完整流程后，还可以通过 Enum.extends 把统一的数据变换方法挂到所有 enum 实例上。",
+  "storybook.stories.CoreCompositionAndExtension.highlights.extends": "Enum.extends",
+  "storybook.stories.CoreCompositionAndExtension.highlights.raw": "raw()",
+  "storybook.stories.CoreCompositionAndExtension.highlights.merge": "合并枚举",
+  "storybook.stories.CoreCompositionAndExtension.highlights.readonly": "只读结构",
+  "storybook.stories.CoreCompositionAndExtension.section.result.title": "组合结果",
+  "storybook.stories.CoreCompositionAndExtension.section.result.description": "利用 raw() 回收原始定义，再重新组合成新的 enum。",
+  "storybook.stories.CoreCompositionAndExtension.field.frozenEnum": "Object.isFrozen(enum)",
+  "storybook.stories.CoreCompositionAndExtension.field.frozenItems": "Object.isFrozen(items)",
+  "storybook.stories.CoreCompositionAndExtension.field.allLabels": "所有 label",
+  "storybook.stories.CoreCompositionAndExtension.field.enumName": "枚举名",
+  "storybook.stories.CoreCompositionAndExtension.section.extension.title": "自定义扩展方法",
+  "storybook.stories.CoreCompositionAndExtension.section.extension.description": "这里通过 Enum.extends 增加了 toBadgeMap()，把一份枚举统一转成适合状态徽标展示的数据结构。",
+  "storybook.stories.CoreCompositionAndExtension.section.extension.codeTitle": "toBadgeMap() 注册代码",
+  "storybook.stories.CoreCompositionAndExtension.sample.draft": "草稿",
+  "storybook.stories.CoreCompositionAndExtension.sample.review": "审核中",
+  "storybook.stories.CoreCompositionAndExtension.sample.published": "已发布",
+  "storybook.stories.CoreCompositionAndExtension.sample.archived": "已归档",
+  "storybook.stories.CoreCompositionAndExtension.sample.fullFlow": "完整内容流程",
   "storybook.stories.CorePatterns.templates.page.title": "全局模板：一次配置，惠及所有枚举",
   "storybook.stories.CorePatterns.templates.page.description": "只需设置一次 Enum.config.templates，之后每个枚举都只需要极简的初始化对象加上 options.name。key 会自动推导，并通过 Enum.localize 完成国际化。",
   "storybook.stories.CorePatterns.templates.mode.flat": "扁平模式 (Flat)",
